Close tarfile after extraction
It's generally not good practice to rely on garbage collection to close
open files, so we should call close() after extraction. I also removed
the blank print lines and instead use the newline character "\n" for
cleaner code. Finally, print is called as a function for Python 3.x
forwards compatibility.